% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_AllDescendants.R
\name{add_AllDescendants}
\alias{add_AllDescendants}
\title{Function for adding AllDescendants field to a ExTree file.}
\usage{
add_AllDescendants(ExTree)
}
\arguments{
\item{ExTree}{A ExTree file obtained from conv_ExTree.}
}
\value{
Return a ExTree file.
}
\description{
Function for adding AllDescendants field to a ExTree file.
}
