% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_ExpreBias.R
\name{add_ExpreBias}
\alias{add_ExpreBias}
\title{Function for adding ExpreBias field during LEU estimation.}
\usage{
add_ExpreBias(ExTree)
}
\arguments{
\item{ExTree}{A ExTree file obtained from conv_ExTree.}
}
\value{
Return a ExTree file.
}
\description{
Function for adding ExpreBias field during LEU estimation.
}
