% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_FilterBiasParent.R
\name{add_FilterBiasParent}
\alias{add_FilterBiasParent}
\title{Function for adding FilterBiasParent field to a ExTree file.}
\usage{
add_FilterBiasParent(ExTree)
}
\arguments{
\item{ExTree}{A ExTree file obtained from conv_ExTree.}
}
\value{
Return a ExTree file.
}
\description{
Function for adding FilterBiasParent field to a ExTree file.
}
