% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_MonoClades.R
\name{add_MonoClades}
\alias{add_MonoClades}
\title{Function for adding MonoClades field to a ExTree file.}
\usage{
add_MonoClades(ExTree)
}
\arguments{
\item{ExTree}{A ExTree file obtained from conv_ExTree.}
}
\value{
Return a ExTree file.
}
\description{
Function for adding MonoClades field to a ExTree file.
}
