% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_TCA_input.R
\name{check_TCA_input}
\alias{check_TCA_input}
\title{Checking the input files for TCA.}
\usage{
check_TCA_input(Tree, Ann)
}
\arguments{
\item{Tree}{A tree file of class "phylo" with node labels (using node ID by default).}

\item{Ann}{A dataframe with TipLabel and TipAnn (tip labels on the tree file and corresponding cell annotations).}
}
\description{
Checking the input files for TCA.
}
