% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conv_ExTree.R
\name{conv_ExTree}
\alias{conv_ExTree}
\title{Convert a tree file and a annotation dataframe to ExTree format.}
\usage{
conv_ExTree(Tree, Ann, ForceFactor = TRUE)
}
\arguments{
\item{Tree}{A tree file of class "phylo" with node labels (using node ID by default).}

\item{Ann}{A dataframe with TipLabel and TipAnn (tip labels on the tree file and corresponding cell annotations).}

\item{ForceFactor}{Force transform TipAnn into a factor (default TRUE).}
}
\value{
Return a ExTree file.
}
\description{
Convert a tree file and a annotation dataframe to ExTree format.
}
\examples{
library("TarCA")
load(system.file("Exemplar","Exemplar_TCA.RData",package = "TarCA"))
tmp.ExTree <- conv_ExTree(Tree = ExemplarData_1$Tree,Ann = ExemplarData_1$Ann)
}
