% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LEU_Estimator.R
\name{LEU_Estimator}
\alias{LEU_Estimator}
\title{Detection of lineage specific expression upregulation.}
\usage{
LEU_Estimator(Tree, Ann, ExTree = NULL)
}
\arguments{
\item{Tree}{A tree file of class "phylo" with node labels (using node ID by default).}

\item{Ann}{A dataframe with TipLabel and TipAnn (tip labels on the tree file and corresponding cell annotations).}

\item{ExTree}{(optional) A ExTree file obtained from conv_ExTree.}
}
\value{
A list containing Tree, Ann, BiasNode, FilterBiasNode, BiasFig, PureNode, PureNode2Organ and EffN. The estimated Np are store in EffN.
}
\description{
Detection of lineage specific expression upregulation.
}
\examples{
library("TarCA")
load(system.file("Exemplar","Exemplar_LEU.RData",package = "TarCA"))
tmp.res <- LEU_Estimator(
Tree = ExemplarData_2$Tree,
Ann = ExemplarData_2$Ann,
Fileout = NULL,
ReturnNp = TRUE
)
}
