% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_MonoClades.R
\name{add_MonoClades}
\alias{add_MonoClades}
\title{Title}
\usage{
add_MonoClades(ExTree)
}
\arguments{
\item{ExTree}{A ExTree file obtained from conv_ExTree.}
}
\value{
Return
}
\description{
Title
}
\examples{
library("TarCA")
Example
}
