% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conv_ExTree.R
\name{conv_ExTree}
\alias{conv_ExTree}
\title{Title}
\usage{
conv_ExTree(Tree, Ann, ForceFactor = TRUE)
}
\arguments{
\item{Tree}{A tree file of class "phylo" with node labels (using node ID by default).}

\item{Ann}{A dataframe with TipLabel and TipAnn (tip labels on the tree file and corresponding cell annotations).}

\item{ForceFactor}{Force transform TipAnn into a factor (default TRUE).}
}
\value{
Return
}
\description{
Title
}
\examples{
library("TarCA")
Example
}
